requirements.txt
cherrypy>=18.8.0
jinja2>=3.1.2
mysql-connector-python>=8.0.33
bcrypt>=4.0.1
matplotlib>=3.7.1 